resourceの名前空間の分離
