+++ lib/constants/model_keywords.js
@@ -1,0 +1,7 @@
+'use strict'
+
+exports.ENTITY_SUFFIX = 'Entity'
+exports.ENTITY_ATTRIBUTE_SUFFIX = 'EntityAttr'
+
+exports.ID_SUFFIX = 'Id'
+
