+++ lib/helpers/builder.js
@@ -10,7 +10,7 @@ const as = (prefix, name) => [ prefix, name ].join('X')
   .replace(/\$/g, '_')
   .replace(/\./g, '_')
 const sortAs = (name) => as('sorting', name)
-const filterAs = (name) => as('filtering', name)
+const filterAs = (name, scope) => as('filtering', name)
 const sortAbs = (name) => name.replace(/^-/, '')
 
 function buildWhereTerm ({ QueryGenerator, filter }) {
@@ -26,19 +26,19 @@ function buildOrderTerm ({ sort }) {
   }).filter(Boolean).join(', ')
 }
 
-function buildSelectSQL ({ filter, sort, limit, offset, orderTerm, whereTerm }) {
+function buildSelectSQL ({ modelName, modelIdName, attributeModelRefName, filter, sort, limit, offset, orderTerm, whereTerm }) {
   const filterKeys = Object.keys(filter || {})
   return `
 SELECT
   \`Entity\`.id as id 
-FROM \`Entities\` AS \`Entity\`
+FROM \`${modelName}\` AS \`Entity\`
 ${filterKeys.map((name) => `
-INNER JOIN \`EntityAttributes\` AS \`${filterAs(name)}\`
-  ON \`Entity\`.\`id\` = \`${filterAs(name)}\`.\`EntityId\` AND \`${filterAs(name)}\`.\`name\` = '${name}'
+INNER JOIN \`${attributeModelRefName}\` AS \`${filterAs(name)}\`
+  ON \`Entity\`.\`id\` = \`${filterAs(name)}\`.\`${modelIdName}\` AND \`${filterAs(name)}\`.\`name\` = '${name}'
 `).join(' ')}
 ${sort.map((name) => `
-LEFT OUTER JOIN \`EntityAttributes\` AS \`${sortAs(sortAbs(name))}\`
-  ON \`Entity\`.\`id\` = \`${sortAs(sortAbs(name))}\`.\`EntityId\` AND \`${sortAs(sortAbs(name))}\`.\`name\` = '${sortAbs(name)}'
+LEFT OUTER JOIN \`${attributeModelRefName}\` AS \`${sortAs(sortAbs(name))}\`
+  ON \`Entity\`.\`id\` = \`${sortAs(sortAbs(name))}\`.\`${modelIdName}\` AND \`${sortAs(sortAbs(name))}\`.\`name\` = '${sortAbs(name)}'
 `).join(' ')}
 WHERE \`Entity\`.\`ResourceId\` = :ResourceId 
 ${whereTerm ? `AND ${whereTerm.trim()}` : ''}    
@@ -52,15 +52,15 @@ ${offset ? `OFFSET ${offset}` : ''}
  * @function buildCountSQL
  * @returns {string}
  */
-function buildCountSQL ({ filter, whereTerm }) {
+function buildCountSQL ({ modelName, modelIdName, attributeModelRefName, filter, whereTerm }) {
   const filterKeys = Object.keys(filter || {})
   return `
 SELECT 
   count(DISTINCT(\`Entity\`.\`id\`)) AS \`count\`
-FROM \`Entities\` AS \`Entity\`
+FROM \`${modelName}\` AS \`Entity\`
 ${filterKeys.map((name) => `
-    INNER JOIN \`EntityAttributes\` AS \`${filterAs(name)}\`
-      ON \`Entity\`.\`id\` = \`${filterAs(name)}\`.\`EntityId\` AND \`${filterAs(name)}\`.\`name\` = '${name}'
+    INNER JOIN \`${attributeModelRefName}\` AS \`${filterAs(name)}\`
+      ON \`Entity\`.\`id\` = \`${filterAs(name)}\`.\`${modelIdName}\` AND \`${filterAs(name)}\`.\`name\` = '${name}'
     `).join(' ')}
 WHERE \`Entity\`.\`ResourceId\` = :ResourceId
 ${whereTerm ? `AND ${whereTerm}` : ''}
