+++ lib/helpers/normalizer.js
