+++ lib/helpers/parser.js
