+++ lib/helpers/serializer.js
