+++ lib/modeling/index.js
@@ -8,9 +8,7 @@
 let d = (module) => module && module.default || module
 
 module.exports = {
-  get builder () { return d(require('./builder')) },
-  get defineModels () { return d(require('./define_models')) },
-  get normalizer () { return d(require('./normalizer')) },
-  get parser () { return d(require('./parser')) },
-  get serializer () { return d(require('./serializer')) }
+  get entityAttributeModel () { return d(require('./entity_attribute_model')) },
+  get entityModel () { return d(require('./entity_model')) },
+  get resourceModel () { return d(require('./resource_model')) }
 }
