+++ lib/sequelize_driver.js
@@ -13,7 +13,7 @@
 const { Driver } = require('clay-driver-base')
 const asleep = require('asleep')
 const co = require('co')
-const { defineModels } = require('./modeling')
+const { resourceModel, entityAttributeModel, entityModel } = require('./modeling')
 const clayId = require('clay-id')
 const clayEntity = require('clay-entity')
 const clayCollection = require('clay-collection')
@@ -21,29 +21,37 @@ const { pageToOffsetLimit } = require('clay-list-pager')
 const Sequelize = require('sequelize')
 const cls = require('continuation-local-storage')
 const clayResourceName = require('clay-resource-name')
-const { LogPrefixes } = require('clay-constants')
 const sgQueue = require('sg-queue')
 const retry = require('retry-as-promised')
+const { pascalcase } = require('stringcase')
 
-let queue = sgQueue()
+const queue = sgQueue()
 
 Sequelize.useCLS(
   cls.createNamespace('clay:driver:sequelize-transaction')
 )
 
-const { DRIVER_PREFIX } = LogPrefixes
-
 /** @lends SequelizeDriver */
 class SequelizeDriver extends Driver {
-
   constructor (database, username, password, options) {
     super()
     const s = this
 
-    let db = new Sequelize(database, username, password, options)
+    const db = new Sequelize(database, username, password, options)
+
+    // TODO model定義の方を直す
+    db.define = ((oritinal) =>
+      // Migration from sequelize v3 to v4
+      function defineFallback (name, attributes, options = {}) {
+        const { classMethods = {}, instanceMethods = {} } = options
+        const defined = oritinal.call(db, name, attributes, Object.assign({}, options))
+        Object.assign(defined, classMethods)
+        Object.assign(defined.prototype, instanceMethods)
+        return defined
+      })(db.define)
+
     s._db = db
-    s._synced = false
-    s._models = defineModels(db)
+    s._models = {}
 
     // TODO Database lockを回避するためにqueuingしているが、あくまで暫定対処。
     // パフォーマンス的に残念、冗長化すると死ぬ等の問題があるので根本対策が必要
@@ -65,22 +73,43 @@ class SequelizeDriver extends Driver {
       )
   }
 
-  getModels () {
+  getEntityModels (resource) {
     const s = this
     return co(function * () {
-      if (!s._synced) {
-        yield s._db.sync()
-        s._synced = true
+      const Resource = yield s.getResourceModel()
+      const prefix = pascalcase(resource.name).replace(/_/g, '')
+      if (!s._models[ prefix ]) {
+        const { _db: db } = s
+        const EntityAttribute = entityAttributeModel({ db, Resource, prefix })
+        const Entity = entityModel({ db, Resource, EntityAttribute, prefix })
+        yield Resource.sync()
+        yield Entity.sync()
+        yield EntityAttribute.sync()
+        s._models[ prefix ] = { Entity, EntityAttribute }
       }
-      return s._models
+      return s._models[ prefix ]
+    })
+  }
+
+  getResourceModel () {
+    const s = this
+    return co(function * () {
+      if (!s._models.Resource) {
+        const { _db: db } = s
+        const Resource = resourceModel({ db })
+        yield Resource.sync()
+        s._models.Resource = Resource
+      }
+      return s._models.Resource
     })
   }
 
   one (resourceName, id) {
     const s = this
     return co(function * () {
-      const { Resource, Entity } = yield s.getModels()
+      const Resource = yield s.getResourceModel()
       const resource = yield Resource.ofName(resourceName)
+      const { Entity } = yield s.getEntityModels(resource)
       let cid = String(id)
       let entity = yield Entity.forOne(resource, cid)
       if (!entity) {
@@ -94,8 +123,9 @@ class SequelizeDriver extends Driver {
   list (resourceName, condition = {}) {
     const s = this
     return co(function * () {
-      const { Resource, Entity } = yield s.getModels()
+      const Resource = yield s.getResourceModel()
       const resource = yield Resource.ofName(resourceName)
+      const { Entity } = yield s.getEntityModels(resource)
       let { filter = {}, page = {}, sort = [] } = condition
       let { limit, offset } = pageToOffsetLimit(page)
       let { rows, count } = yield Entity.forList(resource, { offset, limit, filter, sort })
@@ -114,8 +144,9 @@ class SequelizeDriver extends Driver {
   create (resourceName, attributes) {
     const s = this
     return co(function * () {
-      const { Resource, Entity, EntityAttribute } = yield s.getModels()
+      const Resource = yield s.getResourceModel()
       const resource = yield Resource.ofName(resourceName)
+      const { Entity, EntityAttribute } = yield s.getEntityModels(resource)
       const { id = clayId() } = attributes
       const cid = String(id)
       yield s.transaction(() => co(function * () {
@@ -133,8 +164,9 @@ class SequelizeDriver extends Driver {
   update (resourceName, id, attributes) {
     const s = this
     return co(function * () {
-      const { Resource, Entity, EntityAttribute } = yield s.getModels()
+      const Resource = yield s.getResourceModel()
       const resource = yield Resource.ofName(resourceName)
+      const { Entity, EntityAttribute } = yield s.getEntityModels(resource)
       let cid = String(id)
       const entity = yield Entity.forOne(resource, cid)
       Entity.forOneCache.del(Entity.cacheKeyFor(resource, cid))
@@ -149,8 +181,9 @@ class SequelizeDriver extends Driver {
   destroy (resourceName, id) {
     const s = this
     return co(function * () {
-      const { Resource, Entity } = yield s.getModels()
+      const Resource = yield s.getResourceModel()
       const resource = yield Resource.ofName(resourceName)
+      const { Entity } = yield s.getEntityModels(resource)
       let cid = String(id)
       const entity = yield Entity.forOne(resource, cid)
       if (!entity) {
@@ -167,12 +200,13 @@ class SequelizeDriver extends Driver {
   drop (resourceName) {
     const s = this
     return co(function * () {
-      const { Resource, Entity } = yield s.getModels()
+      const Resource = yield s.getResourceModel()
       Resource.clearCacheForName(resourceName)
       const resource = yield Resource.ofName(resourceName)
       if (!resource) {
         return 0
       }
+      const { Entity } = yield s.getEntityModels(resource)
       return s.transaction(() => co(function * () {
         yield Entity.destroy({
           where: { ResourceId: resource.id }
@@ -186,7 +220,7 @@ class SequelizeDriver extends Driver {
   resources () {
     const s = this
     return co(function * () {
-      const { Resource } = yield s.getModels()
+      const Resource = yield s.getResourceModel()
       const resources = yield Resource.findAll({})
       return resources.map((resource) => {
         const { name, domain } = clayResourceName(resource)
