+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "clay-driver-sequelize",
-  "version": "4.0.11",
+  "version": "4.1.0",
   "description": "Clay driver for Sequelize",
   "main": "lib",
   "browser": "ci/browser",
@@ -38,6 +38,7 @@
     "retry-as-promised": "^2.2.0",
     "sequelize": "^4.3.2",
     "sg-queue": "^1.1.4",
+    "stringcase": "^4.0.0",
     "update": "^0.7.4"
   },
   "devDependencies": {
