--- test/builder_test.js
@@ -1,27 +1,0 @@
-/**
- * Test case for builder.
- * Runs with mocha.
- */
-'use strict'
-
-const builder = require('../lib/modeling/builder.js')
-const assert = require('assert')
-const co = require('co')
-
-describe('builder', function () {
-  this.timeout(3000)
-
-  before(() => co(function * () {
-
-  }))
-
-  after(() => co(function * () {
-
-  }))
-
-  it('Builder', () => co(function * () {
-
-  }))
-})
-
-/* global describe, before, after, it */
