--- test/normalizer_test.js
@@ -1,27 +1,0 @@
-/**
- * Test case for normalizer.
- * Runs with mocha.
- */
-'use strict'
-
-const normalizer = require('../lib/modeling/normalizer.js')
-const assert = require('assert')
-const co = require('co')
-
-describe('normalizer', function () {
-  this.timeout(3000)
-
-  before(() => co(function * () {
-
-  }))
-
-  after(() => co(function * () {
-
-  }))
-
-  it('Normalizer', () => co(function * () {
-
-  }))
-})
-
-/* global describe, before, after, it */
