--- test/parser_test.js
@@ -1,27 +1,0 @@
-/**
- * Test case for parser.
- * Runs with mocha.
- */
-'use strict'
-
-const parser = require('../lib/modeling/parser.js')
-const assert = require('assert')
-const co = require('co')
-
-describe('parser', function () {
-  this.timeout(3000)
-
-  before(() => co(function * () {
-
-  }))
-
-  after(() => co(function * () {
-
-  }))
-
-  it('Parser', () => co(function * () {
-
-  }))
-})
-
-/* global describe, before, after, it */
