--- test/serializer_test.js
@@ -1,27 +1,0 @@
-/**
- * Test case for serializer.
- * Runs with mocha.
- */
-'use strict'
-
-const serializer = require('../lib/modeling/serializer.js')
-const assert = require('assert')
-const co = require('co')
-
-describe('serializer', function () {
-  this.timeout(3000)
-
-  before(() => co(function * () {
-
-  }))
-
-  after(() => co(function * () {
-
-  }))
-
-  it('Serializer', () => co(function * () {
-
-  }))
-})
-
-/* global describe, before, after, it */
